+++ src/acconeer/exptool/app/new/plugin_loader.py
@@ -12,6 +12,7 @@ def load_default_plugins() -> list[PluginSpec]:
     from acconeer.exptool.a121.algo.distance._detector_plugin import DISTANCE_DETECTOR_PLUGIN
     from acconeer.exptool.a121.algo.phase_tracking._plugin import PHASE_TRACKING_PLUGIN
     from acconeer.exptool.a121.algo.presence._detector_plugin import PRESENCE_DETECTOR_PLUGIN
+    from acconeer.exptool.a121.algo.smart_presence._ref_app_plugin import SMART_PRESENCE_PLUGIN
     from acconeer.exptool.a121.algo.sparse_iq._plugin import SPARSE_IQ_PLUGIN
     from acconeer.exptool.a121.algo.touchless_button._plugin import TOUCHLESS_BUTTON_PLUGIN
     from acconeer.exptool.a121.algo.vibration._plugin import VIBRATION_PLUGIN
@@ -25,4 +26,5 @@ def load_default_plugins() -> list[PluginSpec]:
         VIBRATION_PLUGIN,
         BILATERATION_PLUGIN,
         BREATHING_PLUGIN,
+        SMART_PRESENCE_PLUGIN,
     ]